Generic Special Movement
Special movement is now also defined by a "condition", a method which returns true or false. The move will be processed like a normal move, except the condition will also be run. The move is only valid if the condition is true.

This essentially makes special movement as easy as defining a new condition method in the class for a piece.

To complete special movement, we also have to set up actions- code which runs when a special move is executed